requirements.txt
folium==0.14.0
haversine==2.8.0
numpy==1.24.3
pandas==2.0.1
Pillow==9.5.0
Pillow==9.5.0
plotly==5.14.1
streamlit==1.22.0
streamlit_folium==0.11.1
